requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
groq>=0.4.2
pypdf>=3.17.4
pydantic>=2.6.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
